+++ back/src/main/java/com/example/monday/resource/CreateStudent.java
@@ -24,5 +24,5 @@ public class CreateStudent {
     private Semester semester;
 
     @NotNull
-    private float czesne;
+    private Long czesne;
 }
