+++ front/src/main/java/com/example/monday/resource/CreateStudent.java
@@ -22,5 +22,5 @@ public class CreateStudent {
     private Semester semester;
 
     @NotNull
-    private  float czesne;
+    private  Long czesne;
 }
