+++ front/src/main/java/com/example/monday/resource/StudentDto.java
@@ -2,5 +2,5 @@ package com.example.monday.resource;
 
 import java.util.UUID;
 
-public record StudentDto(UUID id, String name, StudentUnit unit, Semester semester, float czesne, Long index) {
+public record StudentDto(UUID id, String name, StudentUnit unit, Semester semester, Long czesne, Long index) {
 }
